refactor: load zip.js asynchronously
