+++ src/options/views/tab-settings/vm-import.vue
@@ -21,6 +21,7 @@
 import { i18n, sendMessage } from '#/common';
 import options from '#/common/options';
 import SettingCheck from '#/common/ui/setting-check';
+import loadZip from '#/common/zip';
 import { showMessage } from '../../utils';
 
 export default {
@@ -76,7 +77,8 @@ function getVMConfig(text) {
 function getVMFile(entry, vmFile) {
   if (!entry.filename.endsWith('.user.js')) return;
   const vm = vmFile || {};
-  return new Promise((resolve) => {
+  return loadZip()
+  .then(zip => new Promise((resolve) => {
     const writer = new zip.TextWriter();
     entry.getData(writer, (text) => {
       const data = { code: text };
@@ -97,11 +99,12 @@ function getVMFile(entry, vmFile) {
       })
       .then(() => resolve(true), () => resolve());
     });
-  });
+  }));
 }
 
 function getVMFiles(entries) {
-  return new Promise((resolve) => {
+  return loadZip()
+  .then(zip => new Promise((resolve) => {
     const data = { entries };
     const i = entries.findIndex(entry => entry.filename && entry.filename.toLowerCase() === 'violentmonkey');
     if (i < 0) {
@@ -114,17 +117,18 @@ function getVMFiles(entries) {
       data.vm = getVMConfig(text);
       resolve(data);
     });
-  });
+  }));
 }
 
 function readZip(file) {
-  return new Promise((resolve, reject) => {
+  return loadZip()
+  .then(zip => new Promise((resolve, reject) => {
     zip.createReader(new zip.BlobReader(file), (res) => {
       res.getEntries((entries) => {
         resolve(entries);
       });
     }, (err) => { reject(err); });
-  });
+  }));
 }
 
 function importData(file) {
